CLIM-1367: Atom1 Moving out what is now called Forecast Summary
